This commit adds all reactor project's apidocs to the search path for external links. It is related to MJAVADOC-97, but only deals with reactor projects.
Only already built reactor projects are added.
The OfflineLink is used to preserve the official site link (using the pom's url with '/apidocs' appended),
and the real location will be the project's javadoc output directory.

Issues/Todo's:
- it is assumed that remote javadoc's are indeed in the project's '/apidocs' uri
- it is assumed that the relative path from the current project to the generated javadocs
  ('target/site/apidoc' or something) is the same for all reactor projects.
- also add links to other, non-reactor dependencies - but configurable.
- when using site:stage, that mojo will replace all URLs in the poms to a huge relative
  path which goes up way too far (bug in pathtool?). So these links may not work
  in a staged site. This however has nothing to do with this commit as far as I can see;
  maybe site:stage is broken...

Fix for the first two: somehow get the javadoc plugin configuration for each reactor project
and read the values there (and interpolate them). Something similar as is done in the eclipse
plugin, which read's the compiler plugin's <source> element, can be done here.

The offline link approach was taken because if you use the real urls then the first
deploy will always fail since the site isn't up yet. This also works fine in offline mode.



git-svn-id: https://svn.apache.org/repos/asf/maven/plugins/trunk@569208 13f79535-47bb-0310-9956-ffa450edef68