[MJAVADOC-620] Integration test for class in top-level package
The first case in which the module info retrieval fails is a JAR
containing a class in the top-level package. This is the originally
reported case in the JIRA ticket MJAVADOC-620.
The FindException has a root cause that says that the usage of
top-level packages is not allowed in modules.

For the test, it is important to build the JAR in a Maven module that
is *not* part of the Reactor. Otherwise, javadoc:aggregate would use
the sources of class Test directly and the bug would not occur.
This has been achieved by *not* adding the test project
maven-MJAVADOC620-jar as a <module> to the main project. Instead,
in invoker.properties, the first project is built using option -f.

Note: The artifactId may not be maven-MJAVADOC-620-jar (with a dash
before the issue number), because then, the module info retrieval
fails for another reason, namely not being able to derive a module
name. This is tested by an upcoming second integration test.

Without the fix, maven-MJAVADOC620-jar-1.0-SNAPSHOT.jar is not added
to the classpath and building the JavaDoc fails, because class Test
is not found.
With the fix, maven-MJAVADOC620-jar-1.0-SNAPSHOT.jar is added
to the classpath and building the JavaDoc succeeds.