[MJAVADOC-620] Integration test for JAR with version without dots
The second case in which the module info retrieval fails is a JAR
with a file name that contains a number *not* followed by a dot.
This is the case we experienced, because we often use 1-SNAPSHOT as
the working version, resulting in a JAR file name like
"foo-1-SNAPSHOT.jar".
The FindException has a root cause that says that the module name
derived from the file name, "foo.1.SNAPSHOT", is invalid because "1"
is not a valid Java identifier.

Compared to the first integration test, the version number of the
test project has been changed from "1.0" to just "1".
To test one failure cause at a time, the test class Test has been
moved from the top-level package into package "somepackage".

Without the fix, maven-MJAVADOC620-jar-1-SNAPSHOT.jar is not added
to the classpath and building the JavaDoc fails, because class Test
is not found.
With the fix, maven-MJAVADOC620-jar-1-SNAPSHOT.jar is added
to the classpath and building the JavaDoc succeeds.