[MJAVADOC-617] Normalize module path so that '..' in the path are resolved (#27)
* MJAVADOC-617 Normalize module path so that '..' in the path are resolved

On line 2403 a path is removed from a map. The given path may not be normalized but the pathes in the map are. This leads to the fact that the project for the given path is not found. And therefore no javadoc for the project is generated.
By normalizing the module path on line 2396 the code works event if the module path has '..' in it.

* MJAVADOC-617 Add regression test

* MJAVADOC-617 Simplify code as suggested by review

* MJAVADOC-617 Fix checkstyle problems

* MJAVADOC-617 Refactor duplicated code into abstract sub classes

- Refactor duplicated code into common new
AbstractAggregateMavenProjectStub and
AbstractAggregateChildMavenProjectStub classes
- Fix toLowerCase without Locale
- Simplify method getOverviewSummary and make it static

* MJAVADOC-617 improve test assertions