[MJAVADOC-770] Implement legacy mode for Java 9+
A lot of existing projects that don't use module descriptors but only
automatic module naming have moved past Java 8. However, when trying
to build docs, the javadoc plugin assumes that the project is
modularized and starts using the module path which leads to problems
with documentation generation.

This change introduces a new configuration setting `<legacyMode>`
which is turned off by default. When turning on, the javadoc plugin
will continue to construct just a classpath (and not a module path)
for the javadoc tool, even if the jdk version is 9 or greater.

Includes integration test and documentation.